METEO_KAFKA_STREAMLIT/requirements.txt
altair==5.3.0
attrs==23.2.0
backports.zoneinfo==0.2.1
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
importlib-metadata==7.1.0
importlib-resources==6.4.0
Jinja2==3.1.4
jsonschema==4.22.0
kafka-python==2.0.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==1.26.4
packaging==24.0
pandas==2.2.2
Pillow==10.3.0
pkgutil_resolve_name==1.3.10
protobuf==5.27.1
pyarrow==16.1.0
pydeck==0.9.0b1
Pygments==2.18.0
Pympler==1.0.1
pyrsistent==0.20.0
python-dateutil==2.9.0
python-dotenv==1.0.1
pytz==2023.3.post1
pytz-deprecation-shim==0.1.0.post0
requests==2.32.3
rich==13.7.1
six==1.16.0
smmap==6.0.0
streamlit==1.35.0
tenacity==8.3.0
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
typing_extensions==4.12.1
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
validators==0.28.3
zipp==3.19.2
kafka-python-ng==2.2.2
kafka-python   
streamlit
python-dotenv
plotly